patch/user: Introduced CacheLocationFixer
CacheLocationFixer allows multi-installation users to cleanly
separate their game installation cache folders.

The patch looks which folder is the game executable in and forces
the game to direct anything to do with cache there as opposed to
`%LocalAppData%\Growtopia` that the game normally uses.

Most notable usecase of this feature is for people that want to
play regular Growtopia and also keep an older game version around
so they can use osgt-qol.

NOTE: This currently does not separate save.dat. More research has
to be done to determine if that is feasible yet.